Simplified readability; added note
Name of item 3 was not simple enough. Always striving for simplicity
here. Added note on details for contributor - didn’t want to include it
in the numbered steps as it unnecessarily complicated the steps from a
beginners POV.

Clarified top of technical file because that intro was leftover from
before when it was inside the CONTRIB