Reordered to ensure everything is read
Reordered the top of the CONTRIB to be more easily read. I was
concerned that visitors might not read the ‘Getting Started’ section,
so I put that up front and then explained the they can click on the 4
ways to contribute in order to learn more about each of them.

Also updated README to match CONTRIB