Rename files so we can start them anew
Because these files/patterns shouldn’t actually really be in master
yet, and we want to create new PR’s for each pattern, if we don’t
rename the files, then the newly created PR’s will only show small
changes made to the existing files, and therefore be difficult to make
reviews on. I tested this and reviewers could only see small changes
made. SO, if we rename the files, and start a new on the files, with
new PR’s on what looks like new files, this should allow full reviews.
We will lose some history on two files and this is a small downside.