This adds the maturity model as proposed in #116
This pull requests converts the maturity model proposed in #116 into pattern
format. It converts the table shared in the issue into something that hopefully
is easier to review - but also easier to reference. I followed the format of the
Apache Maturity Model
https://community.apache.org/apache-way/apache-project-maturity-model.html for
formatting the individual points. Links (anchors in the page to individual
points) might make referencing points more precise but are missing at this
point.

I converted the table as is - we can tidy up and clarify going forward. At least
IMHO there are still a few points that are debatable.