Update the link to the public page I referred to
Hi reviewer, I've added the link here, but I'm sure if placing the link on the name is good or not. Please feel free to adjust or let me do it if needed. Thanks.
(Actually, this is my first time contributing to public repositories; please let me know if I violate some rules.)