Adding Know Instance from DAZN
This blog post by DAZN already contains more info about their RFC process than what we normally need for an org to be added as a Known Instance to a pattern.
https://medium.com/dazn-tech/building-a-dx-team-lessons-learned-4a66446088bc

I am opening a PR for this, containing a summary of what was in the blog post. 
Will then reach out to DAZN (kind of a request-for-comments) to see if they want to add anything else about their use of RFCs.